Imagine permalink now waits for model for screenshot
